tests/wpt/web-platform-tests/css/css-animations/parsing/animation-play-state-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: getComputedValue().animationPlayState</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-play-state">
<meta name="assert" content="animation-play-state computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("animation-play-state", "running, paused");
</script>
</body>
</html>
